test: integrate GoogleTest and establish core library structure
overview:
integrated GoogleTest (GTest) via CPM, introduced csurfer_lib to
separate browser logic from the main entry point, and migrated existing
manual tests to the GTest framework.

why:
following Refactoring Rule 10, every refactored module needs GTest.
the previous testing setup was manual and didn't support the leaf-to-core
migration safely. extracting csurfer_lib allows us to link the same
production code into both the main executable and various unit test
binaries without duplicating file lists in CMake.

how:
1. added GTest dependency to CMakeLists.txt using CPM.
2. defined csurfer_lib static library containing all src/ logic.
3. created unit-tests/config/test_config.cpp to verify config relationships.
4. converted unit-tests/test_url_origin.cpp to unit-tests/url/test_url.cpp.
5. updated scripts/test/unit.sh to run tests via CTest.

outcome:
automated testing is now part of the build pipeline. we have a scalable
structure for adding tests as we refactor each module. all 2 current
test suites pass.
